// src/types/install-manifest.ts
// Interface for the install manifest
export interface InstallManifest {
    FormatVersion: number;
    bIsIncompleteInstall: boolean;
    LaunchCommand: string;
    LaunchExecutable: string;
    ManifestLocation: string;
    ManifestHash: string;
    bIsApplication: boolean;
    bIsExecutable: boolean;
    bIsManaged: boolean;
    bNeedsValidation: boolean;
    bRequiresAuth: boolean;
    bAllowMultipleInstances: boolean;
    bCanRunOffline: boolean;
    bAllowUriCmdArgs: boolean;
    bLaunchElevated: boolean;
    BaseURLs: string[];
    BuildLabel: string;
    AppCategories: string[];
    ChunkDbs: any[];
    CompatibleApps: any[];
    DisplayName: string;
    InstallationGuid: string;
    InstallLocation: string;
    InstallSessionId: string;
    InstallTags: string[];
    InstallComponents: string[];
    HostInstallationGuid: string;
    PrereqIds: string[];
    PrereqSHA1Hash: string;
    LastPrereqSucceededSHA1Hash: string;
    StagingLocation: string;
    TechnicalType: string;
    VaultThumbnailUrl: string;
    VaultTitleText: string;
    InstallSize: number;
    MainWindowProcessName: string;
    ProcessNames: string[];
    BackgroundProcessNames: string[];
    IgnoredProcessNames: string[];
    DlcProcessNames: string[];
    MandatoryAppFolderName: string;
    OwnershipToken: string;
    SidecarConfigRevision: number;
    CatalogNamespace: string;
    CatalogItemId: string;
    AppName: string;
    AppVersionString: string;
    MainGameCatalogNamespace: string;
    MainGameCatalogItemId: string;
    MainGameAppName: string;
    AllowedUriEnvVars: string[];
  }